backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
email-validator>=2.0.0
python-dotenv>=1.0.0
openai>=1.54.0
requests>=2.32.0
google-auth>=2.34.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.150.0
numpy>=1.26.0
scikit-learn>=1.6.0
pandas>=2.2.0
sqlalchemy>=2.0.36
python-multipart>=0.0.12
aiohttp>=3.11.0
pydantic-settings>=2.6.0
setuptools>=68.0.0

